% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DacianGheorghe.R
\name{DacianGheorghe}
\alias{DacianGheorghe}
\title{DacianGheorghe}
\usage{
DacianGheorghe()
}
\description{
\code{?DacianGheorghe} contains notes on various functions.\cr
Last update: 06/09/2023 14:30 \cr
Functions in DacianGheorghe Package
\strong{DacianGheorghe} \cr
\strong{go} \cr
\strong{table_} \cr
\strong{rquotes} \cr
}
\details{

}
\section{Package info}{


\emph{---copy text in notepad ---}\cr
Location of source code: \emph{C:\\Users\\DacianGheorghe\\OneDrive - UGent\\Misc\\R programming\\R packages\\DacianGheorgheRPackage} \cr
To update DacianGheorghe package run \code{library(roxygen2); roxygenise()} in DacianGheorgheRPackage.Rproj and upload folder to github \url{https://github.com/daciangheorghe/DacianGheorgheRPackage}. \cr
Run \code{devtools::install_github("daciangheorghe/DacianGheorgheRPackage")} anywhere else to reinstall package.
}

